Determine RPM automatically in addition to shooter angle
